+++ 0x1A-sorting_algorithms/sort.h
@@ -29,7 +29,7 @@ void bubble_sort(int *array, size_t size);
 
 /* insertion sort */
 void insertion_sort_list(listint_t **list);
-size_t len_linkedList(listint_t *head);
+unsigned int len_linkedList(listint_t *head);
 void swap_insert(listint_t *pn, listint_t *tmp, listint_t *nn, listint_t **l);
 
 /* selection sort */
